Move logic FileInfo to IndexerBase. Improve title pre-processing. Delete redundant code.
